requirements.txt
pydantic
openai
motor
python-dotenv